Reworked discord sync + notif handling
- Added tag to integration mapping so we can define one or multiple integrations linked to a tag that will filter which notification integration is allowed for a specific tag (ie discord webhook)

+++ src/Discord/DiscordAction.php
@@ -3,6 +3,7 @@
 namespace Seat\Kassie\Calendar\Discord;
 
 use Seat\Kassie\Calendar\Models\Operation;
+use Seat\Services\Exceptions\SettingException;
 
 class DiscordAction
 {
@@ -56,9 +57,11 @@ class DiscordAction
         $event = GuildEvent::fromOperation($operation);
         $event->status = 4; // CANCELLED
 
-        DiscordClient::modifyGuildEvent($operation->discord_guild_event_id, [
-            'status' => 4
-        ]);
+        if ($operation->discord_guild_event_id) {
+            DiscordClient::modifyGuildEvent($operation->discord_guild_event_id, [
+                'status' => 4
+            ]);
+        }
     }
 
     /**
@@ -68,7 +71,9 @@ class DiscordAction
     {
         logger()->debug("Activated operation, recreating guild event on discord");
         // Operation was cancelled and is being reactivated
-        DiscordClient::deleteGuildEvent($operation->discord_guild_event_id);
+        if ($operation->discord_guild_event_id) {
+            DiscordClient::deleteGuildEvent($operation->discord_guild_event_id);
+        }
         // We recreate the event because discord does not allow a transition from CANCELED state
         $this->createGuildEvent($operation);
     }
@@ -78,10 +83,12 @@ class DiscordAction
      */
     public function endGuildEvent(Operation $operation): void
     {
-        logger()->debug("Ending guild event by settings status 3 = COMPLETED");
-        DiscordClient::modifyGuildEvent($operation->discord_guild_event_id, [
-            'status' => 3 // COMPLETED
-        ]);
+        if ($operation->discord_guild_event_id) {
+            logger()->debug("Ending guild event by settings status 3 = COMPLETED");
+            DiscordClient::modifyGuildEvent($operation->discord_guild_event_id, [
+                'status' => 3 // COMPLETED
+            ]);
+        }
     }
 
     /**
@@ -89,9 +96,11 @@ class DiscordAction
      */
     public function updateGuildEvent(Operation $operation): void
     {
-        logger()->debug("Updating guild event");
-        $event = GuildEvent::fromOperation($operation);
-        DiscordClient::modifyGuildEvent($operation->discord_guild_event_id, $event->toArray());
+        if ($operation->discord_guild_event_id) {
+            logger()->debug("Updating guild event");
+            $event = GuildEvent::fromOperation($operation);
+            DiscordClient::modifyGuildEvent($operation->discord_guild_event_id, $event->toArray());
+        }
     }
 
     /**
@@ -118,6 +127,22 @@ class DiscordAction
             logger()->debug("Guild event has been deleted !");
         } else {
             logger()->info("No guild event to delete on discord because the operation does not have a guild event id");
+        }
+    }
+
+    public static function syncWithDiscord($actionType, $operation): void
+    {
+        try {
+            if (setting('kassie.calendar.discord_integration', true)) {
+                (new self())
+                    ->setType($actionType)
+                    ->execute($operation);
+                logger()->debug("call discord action with type $actionType on operation $operation->title");
+            } else {
+                logger()->debug("Discord integration is not activated");
+            }
+        } catch (DiscordActionException|SettingException $e) {
+            logger()->error("Error guild event on discord " . $e->getMessage());
         }
     }
 }
